Move SearchSuggest to Locations view folder
That's the only place it's used.
